+++ @xen-orchestra/web-core/lib/constants.ts
@@ -1,0 +1,3 @@
+export const XCP_LINKS = {
+  GUEST_TOOLS: 'https://docs.xcp-ng.org/vms/#guest-tools',
+}